pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "academic_metrics"
version = "0.1.0-beta"
description = "A tool to extract and format academic data from Web of Science and Crossref"
authors = [
    { name = "Spencer Presley", email = "spencerpresley96@gmail.com" },
    { name = "Cole Barbes", email = "cbarbes1@gulls.salisbury.edu" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp==3.10.8",
    "beautifulsoup4==4.12.3",
    "langchain==0.3.7",
    "langchain_anthropic==0.3.0", 
    "langchain_core==0.3.19",
    "langchain_google_genai==2.0.5",
    "langchain_openai==0.2.9",
    "openai==1.55.0",
    "pydantic==2.9.2",
    "pylatexenc==2.10",
    "pymongo==4.10.1",
    "python-dotenv==1.0.1",
    "selenium==4.26.1",
    "Unidecode==1.3.8",
    "webdriver_manager==4.0.2"
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "piccolo-theme",
    "sphinx-autodoc-typehints",
    "pydata-sphinx-theme",
]

[project.urls]
"Homepage" = "https://github.com/SpencerPresley/COSC425-DATA"
"Bug Tracker" = "https://github.com/SpencerPresley/COSC425-DATA/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["academic_metrics*"]
namespaces = false
